gpu_poker_cfr/testing/openspiel_compare.py
"""
OpenSpiel comparison utilities.

Validates our CFR implementation against OpenSpiel's reference implementation.
"""

import numpy as np
from typing import Dict, Tuple, Optional

# OpenSpiel is optional
try:
    import pyspiel
    from open_spiel.python.algorithms import cfr
    OPENSPIEL_AVAILABLE = True
except ImportError:
    OPENSPIEL_AVAILABLE = False
    pyspiel = None
    cfr = None


def is_openspiel_available() -> bool:
    """Check if OpenSpiel is available."""
    return OPENSPIEL_AVAILABLE


def run_openspiel_cfr(iterations: int = 1000) -> Dict[str, np.ndarray]:
    """
    Run OpenSpiel CFR on Kuhn poker.

    Args:
        iterations: Number of CFR iterations

    Returns:
        Dictionary mapping infoset string to strategy array
    """
    if not OPENSPIEL_AVAILABLE:
        raise ImportError("OpenSpiel not available. Install with: pip install open_spiel")

    game = pyspiel.load_game("kuhn_poker")
    cfr_solver = cfr.CFRSolver(game)

    for _ in range(iterations):
        cfr_solver.evaluate_and_update_policy()

    # Extract average strategy
    average_policy = cfr_solver.average_policy()

    # Convert to dictionary: infoset_string -> [action_probs]
    strategies = {}

    # Iterate through all information states
    for player in range(2):
        for state_string, action_probs in average_policy.action_probability_array.items():
            # action_probs is already a dict of action -> probability
            # Convert to ordered array
            if state_string.startswith(str(player)):
                # Get number of legal actions
                num_actions = len(action_probs)
                probs = np.zeros(num_actions, dtype=np.float32)
                for action, prob in action_probs.items():
                    probs[action] = prob
                strategies[state_string] = probs

    return strategies


def map_our_infoset_to_openspiel(infoset_key: str, player: int) -> str:
    """
    Map our infoset key format to OpenSpiel format.

    Our format: "J:", "J:b", "K:bc", etc.
    OpenSpiel format: "0", "0pb", "1", "1b", etc.

    Args:
        infoset_key: Our infoset key (e.g., "J:b")
        player: Player index (0 or 1)

    Returns:
        OpenSpiel infoset string
    """
    # Our format: "CARD:HISTORY"
    # OpenSpiel format: "PLAYER" followed by history

    parts = infoset_key.split(":")
    card = parts[0]
    history = parts[1] if len(parts) > 1 else ""

    # Card mapping: J=0, Q=1, K=2 in terms of rank
    # But OpenSpiel uses different representation

    # OpenSpiel infoset strings look like:
    # "0" - player 0, no actions yet
    # "0pb" - player 0, after pass and bet
    # "1b" - player 1, after bet

    # Actually, OpenSpiel uses different format
    # Let's just use action comparison

    # For now, return a placeholder
    # Real implementation would need to analyze OpenSpiel's state encoding
    return f"{player}{history}"


def compare_strategies(
    our_strategy: Dict[str, np.ndarray],
    openspiel_strategy: Dict[str, np.ndarray],
    tolerance: float = 0.1
) -> Tuple[bool, str]:
    """
    Compare strategies from our solver and OpenSpiel.

    Args:
        our_strategy: Dict of infoset_key -> probs from our solver
        openspiel_strategy: Dict of infoset_string -> probs from OpenSpiel
        tolerance: Maximum allowed difference per action probability

    Returns:
        (match, report): Whether strategies match and detailed report
    """
    report_lines = ["Strategy Comparison Report", "=" * 50]

    all_match = True

    for our_key, our_probs in our_strategy.items():
        # Try to find matching OpenSpiel infoset
        # This requires understanding the mapping between formats

        # For now, just report our strategies
        report_lines.append(f"\nOurs [{our_key}]: {our_probs}")

    report_lines.append("\n" + "-" * 50)

    for os_key, os_probs in openspiel_strategy.items():
        report_lines.append(f"OpenSpiel [{os_key}]: {os_probs}")

    report = "\n".join(report_lines)
    return all_match, report


def run_comparison(our_iterations: int = 5000, openspiel_iterations: int = 5000) -> str:
    """
    Run full comparison between our solver and OpenSpiel.

    Args:
        our_iterations: Iterations for our solver
        openspiel_iterations: Iterations for OpenSpiel

    Returns:
        Comparison report string
    """
    from gpu_poker_cfr.games.kuhn import KuhnPoker
    from gpu_poker_cfr.solvers.vanilla import VanillaCFR

    report_lines = [
        "GPU-CFR vs OpenSpiel Comparison",
        "=" * 60,
        ""
    ]

    # Run our solver
    report_lines.append(f"Running our VanillaCFR for {our_iterations} iterations...")
    solver = VanillaCFR(KuhnPoker())
    solver.solve(iterations=our_iterations)
    our_expl = solver.exploitability()
    report_lines.append(f"Our exploitability: {our_expl:.6f}")
    report_lines.append("")

    # Report our strategy
    report_lines.append("Our Strategy:")
    report_lines.append("-" * 40)
    for h_idx in range(solver.matrices.num_infosets):
        name = solver.get_infoset_name(h_idx)
        strategy = solver.get_strategy_for_infoset(h_idx)
        player = solver._infoset_player[h_idx]
        infoset = solver._tree.infosets[h_idx]
        actions = [a.name for a in infoset.actions]
        probs_str = ", ".join(f"{a}={p:.3f}" for a, p in zip(actions, strategy))
        report_lines.append(f"  P{player+1} [{name}]: {probs_str}")

    report_lines.append("")

    # Run OpenSpiel if available
    if OPENSPIEL_AVAILABLE:
        report_lines.append(f"\nRunning OpenSpiel CFR for {openspiel_iterations} iterations...")
        try:
            os_strategies = run_openspiel_cfr(openspiel_iterations)
            report_lines.append("OpenSpiel Strategy:")
            report_lines.append("-" * 40)
            for key, probs in sorted(os_strategies.items()):
                report_lines.append(f"  [{key}]: {probs}")
        except Exception as e:
            report_lines.append(f"OpenSpiel error: {e}")
    else:
        report_lines.append("\nOpenSpiel not available. Skipping comparison.")
        report_lines.append("Install with: pip install open_spiel")

    report_lines.append("")
    report_lines.append("=" * 60)
    report_lines.append("Comparison complete.")

    return "\n".join(report_lines)


def validate_against_known_nash() -> Tuple[bool, str]:
    """
    Validate our solver against known Kuhn poker Nash equilibrium.

    The Nash equilibrium for Kuhn poker is well-known:
    - P1 with J: bet with probability α ∈ [0, 1/3]
    - P1 with Q: check always
    - P1 with K: bet always
    - P2 with J facing bet: fold always
    - P2 with Q facing bet: call with probability 1/3 + α
    - P2 with K facing bet: call always
    - Various responses to checks...

    Returns:
        (valid, report): Whether strategy is close to Nash and detailed report
    """
    from gpu_poker_cfr.games.kuhn import KuhnPoker
    from gpu_poker_cfr.solvers.vanilla import VanillaCFR

    solver = VanillaCFR(KuhnPoker())
    solver.solve(iterations=10000)

    report_lines = [
        "Nash Equilibrium Validation",
        "=" * 50,
        ""
    ]

    valid = True
    expl = solver.exploitability()
    report_lines.append(f"Exploitability: {expl:.6f}")

    if expl > 0.05:
        valid = False
        report_lines.append("WARNING: Exploitability too high!")

    report_lines.append("")
    report_lines.append("Key Strategy Checks:")
    report_lines.append("-" * 40)

    # Check key Nash properties
    checks = [
        ("J:", 1, 0.0, 0.4, "P1 Jack bet frequency"),
        ("K:", 1, 0.7, 1.0, "P1 King bet frequency"),
        ("J:b", 0, 0.95, 1.0, "P2 Jack fold vs bet"),
        ("K:b", 1, 0.9, 1.0, "P2 King call vs bet"),
    ]

    for infoset_key, action_idx, low, high, description in checks:
        for h_idx in range(solver.matrices.num_infosets):
            if solver.get_infoset_name(h_idx) == infoset_key:
                strategy = solver.get_strategy_for_infoset(h_idx)
                prob = strategy[action_idx]

                status = "✓" if low <= prob <= high else "✗"
                if not (low <= prob <= high):
                    valid = False

                report_lines.append(
                    f"  {status} {description}: {prob:.3f} (expected {low:.2f}-{high:.2f})"
                )
                break

    report_lines.append("")
    report_lines.append("=" * 50)
    report_lines.append(f"Validation: {'PASSED' if valid else 'FAILED'}")

    return valid, "\n".join(report_lines)
