pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gpu_poker_cfr"
version = "0.1.0"
description = "GPU-accelerated Counterfactual Regret Minimization for poker"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}

dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "cupy-cuda11x>=11.0.0",  # For GTX 16-series (CUDA 11)
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-benchmark>=4.0.0",
    "grimp>=3.0",
    "import-linter>=2.0",
]
openspiel = [
    "open_spiel>=1.3",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["gpu_poker_cfr*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.importlinter]
root_package = "gpu_poker_cfr"

[[tool.importlinter.contracts]]
name = "Layer architecture"
type = "layers"
layers = [
    "gpu_poker_cfr.solvers",
    "gpu_poker_cfr.engine",
    "gpu_poker_cfr.matrix",
    "gpu_poker_cfr.games",
]
# Each layer can only import from layers below it
# solvers -> engine -> matrix -> games
